third_party/WebKit/LayoutTests/fast/scrolling/scroll-element-into-view-expected.txt

Tests that scrollable areas are scrolled to the correct location when an element is focused.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS container.scrollLeft is 10
PASS container.scrollTop is 10
PASS container.scrollLeft is 690
PASS container.scrollTop is 690
PASS container.scrollLeft is 20
PASS container.scrollTop is 20
PASS container.scrollLeft is 680
PASS container.scrollTop is 680
PASS container.scrollLeft is 10
PASS container.scrollTop is 10
PASS container.scrollLeft is 10
PASS container.scrollTop is 10
PASS container.scrollLeft is 20
PASS container.scrollTop is 20
PASS container.scrollLeft is 20
PASS container.scrollTop is 20
PASS container.scrollLeft is 20
PASS container.scrollTop is 20
PASS container.scrollLeft is 680
PASS container.scrollTop is 680
PASS container.scrollLeft is 30
PASS container.scrollTop is 30
PASS container.scrollLeft is 670
PASS container.scrollTop is 670
PASS successfullyParsed is true

TEST COMPLETE
